Fix some -Wsign-conversion warnings

+++ SFconv/ushort_chartraits.h
@@ -51,7 +51,7 @@ namespace std
       { 
         const char_type* __p = __s; 
         while (*__p) ++__p; 
-        return (__p - __s); 
+        return static_cast<size_t>(__p - __s);
       }
 
       static const char_type*
